Changement dans le DDL Ajout de données d'init Ajout API Géolocalisation google (volez pas la clé svp :p) Premières méthodes du service Trajet

+++ carpooling-domaine/src/main/java/edu/fst/m2/ipii/carpooling/domaine/bo/Adresse.java
@@ -38,10 +38,10 @@ public class Adresse implements Serializable {
 	private String codePostal;
 	
 	@Column(name="Latitude", nullable=true)	
-	private Float latitude;
+	private Double latitude;
 	
 	@Column(name="Longitude", nullable=true)	
-	private Float longitude;
+	private Double longitude;
 	
 	@Column(name="Pays", nullable=true, length=255)	
 	private String pays;
@@ -82,27 +82,27 @@ public class Adresse implements Serializable {
 		return codePostal;
 	}
 	
-	public void setLatitude(float value) {
-		setLatitude(new Float(value));
+	public void setLatitude(double value) {
+		setLatitude(new Double(value));
 	}
 	
-	public void setLatitude(Float value) {
+	public void setLatitude(Double value) {
 		this.latitude = value;
 	}
 	
-	public Float getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 	
-	public void setLongitude(float value) {
-		setLongitude(new Float(value));
+	public void setLongitude(double value) {
+		setLongitude(new Double(value));
 	}
 	
-	public void setLongitude(Float value) {
+	public void setLongitude(Double value) {
 		this.longitude = value;
 	}
 	
-	public Float getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
